+++ CHANGELOG.md
@@ -22,6 +22,7 @@
   - Allow find definition of java class usages where definition comes from clojure, like defrecords.
   - Fix: wait for rename to apply before allowing another rename, to ensure suggested name is correct. #1270
   - Process requests in parallel, to prevent typing lag and other performance problems introduced during migration away from lsp4j. #1240
+  - Fix: Avoid wrong ns require after `Create ns and require` code-action/command.
 
 - API/CLI
   - Fix missing diagnostics when `--project-root` is different than current directory. #1245